backend/requirements.txt
# Django dependencies
Django>=5.0,<5.1
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0
django-environ>=0.11,<1.0

# Database
psycopg2-binary>=2.9,<3.0

# Channels (WebSocket)
channels>=4.0,<5.0
channels-redis>=4.1,<5.0
daphne>=4.0,<5.0

# Celery (Background tasks)
celery>=5.3,<6.0
redis>=5.0,<6.0

# Face Recognition
face_recognition>=1.3,<2.0
numpy>=1.26
Pillow>=10.0,<11.0

# Push Notifications
firebase-admin>=6.2,<7.0

# Utilities
python-dateutil>=2.8,<3.0
requests>=2.31,<3.0

# Development
black>=23.0
flake8>=6.0
pytest-django>=4.5
